Kernel Driver: Fix spurious buffer drop warning
The issue occurs when a running capture in continuous capture mode is
terminated, then the wait on the last buffer read is interrupted due to
a signal rather than being woken up from an interrupt.

Add a check to see if the wait was interrupted in beaglelogic_f_read, if the
wait was indeed interrupted, then return -ERESTARTSYS. This fixes the issue.

Signed-off-by: Kumar Abhishek <abhishek@theembeddedkitchen.net>